Inventory and item
Added Interactable, Item, Inventory, InventorySlot and itemObject class.
Added An inventory to player
Added an test item and a lifepotion class.
Added a label to skillSlot which specify the key you need to use it.
Handle the case where the skillslot Key was displayed on skill to chose during in the Skillchoice UI
Handle the case where the player click on a skillslot and then on an other one