databases and enemy AI
Database are now save on a json file and loaded when needed.
Added a JsonWrapper class which wrap the list of element which need to be saved
Enemy now use abilityID rather than a list of ability. The abilities are loaded from the database in the constructor
Loot now have a itemID attribut which handle the loading of the item in the database (this allow mainly to get the up to date item).
Added a AbilityUsageFrequence Enum which allow to pair ability with a usage frenquency
Moved the ability base list in a static class callded AbilityList
Modified some attribut to be public so they can be serialized by the JsonUtility class (Abilityattributs, stats etc)
Moved the parent attributs in the initSlot function of the UIInventorySlot class so it's always called. Its was previously only called the item was not null