DatabaseModel and ResourceList
Added databadeModel for enemy, item, ability and itemEffectAndValue. Thos class only keep serializable value that will later be used to create non serializable instance. This allow to save in the json non serializable datas.
Added DatabaseResourcesList which is a ScriptableObject holding every object used by databaseItems and a unique id (GUID) to retrived them from the list when needed.
Added a custom editor for DatabaseResourcesList allowing to edit, delete or search the objects.