Added node allocation and effect to player
Added PlayerAllocated node which store the allocated nodes GUIDs and level
Player stat now reset list of start on init event if it's initialized with an another player stat. This will allow to be sure there is no stat from equipement or node that no longer exist or have been modified by an update. Each stat the player have will be added on load.
Added a delegate in UIPassiveNode allowing UIPassiveTree To know when there is a change in a node allocation.
REnamed UIPassiveSkillTree to UIPassiveTree
Added mask to passivetree
Added a constant size to passivetree
PassiveTree now derive from Base UI
Added a zoom function in UIPassiveTree
Added Movement to UIPassiveTree
Added IBeginDragHandler in base ui
Base ui now init a dragingOffset to move item with drag around the mouse and not from the center of the target