Localization misc
Added possibility to edit key in key list
Text for a language is now order by type in the editor. The type is the last part of the key after the "_"
Added localization string for most of the current value
Modified Effect and stat description. They now use a {} value with a text describing it's use instead of numbers.
Added a sentenceFormat function as an extension of string to return a lowercase string with an upper case as first letter