BaseLocalization
Added baselocalizationData in localizationText which is english data loaded from the json. This localizationData is used to get all the keys and get the english value for the new languages. This allow to not load english language every time we need to get some information.
Category in localizationTextEditorWindows can now be folded